requirements.txt
langchain-astradb
langchain 
langchain-openai
datasets
pypdf
python-dotenv
flask
together
langchain_ai21
langchain-community
langchain-qdrant
langchain-google-genai

# -e .